requirements.txt
flask==3.0.3
flask-cors==4.0.0
google-generativeai==0.7.2
python-dotenv==1.0.1